moved pugixml, noted external libraries
