+++ src/LSST/M1M3/SS/Settings/AliasApplicationSettings.cpp
@@ -6,7 +6,7 @@
  */
 
 #include <AliasApplicationSettings.h>
-#include <pugixml.hpp>
+#include <pugixml/pugixml.hpp>
 #include <XMLDocLoad.h>
 #include <boost/tokenizer.hpp>
