+++ src/pugixml/pugixml.cpp
